fix more tests

+++ evalml/pipelines/time_series_classification_pipelines.py
@@ -121,7 +121,7 @@ class TimeSeriesClassificationPipeline(TimeSeriesPipelineBase, ClassificationPip
         predictions = self._estimator_predict(features, y)
         predictions.index = y.index
         # import pdb; pdb.set_trace()
-        predictions = self.inverse_transform(predictions.astype(int))
+        predictions = self.inverse_transform(predictions)
         # predictions = predictions.astype(int)
         predictions = pd.Series(predictions, name=self.input_target_name)
